Fixed a bug
Events were stacking together so it was possible to click on yes for one plug and to toogle all the plugs you've previously openned the popup.

To be more clear, when we click on a row a popup appear, JS add event 'yes' (to toogle) and event 'no' (to close the popup). But the popup is just hidden and because we did not purge the events they were stacking if we opens another popup.

Hence if we clicked on 'Plug 1' then cancels and clicked on 'Plug 2' then yes. Both were toogled because the event we added on button 'yes' were kept.